requirements.txt
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
aiosqlite==0.21.0
anyio==4.8.0
asgiref==3.8.1
attrs==25.1.0
beautifulsoup4==4.13.3
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
docstring_parser==0.16
duckdb==1.2.0
Faker==36.1.0
fastapi==0.68.2
frozenlist==1.5.0
gitdb==4.0.12
GitPython==3.1.44
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
joblib==1.4.2
Markdown==3.7
markdown2==2.5.3
multidict==6.1.0
numpy==2.2.2
pandas==2.2.3
pillow==11.1.0
prettier==0.0.7
propcache==0.2.1
pydantic==1.10.21
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-magic==0.4.27
python-multipart==0.0.5
python-slugify==8.0.4
pytz==2025.1
requests==2.32.3
scikit-learn==1.6.1
scipy==1.15.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.38
starlette==0.14.2
text-unidecode==1.3
threadpoolctl==3.5.0
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
uv==0.5.31
uvicorn==0.15.0
yarl==1.18.3
